+++ components/admin/ActionButtons/ActionButtons.js
@@ -3,10 +3,12 @@ import PropTypes from 'prop-types';
 import { Button, Confirm, Modal } from 'semantic-ui-react';
 
 import ConfirmModalContent from 'components/admin/ConfirmModalContent/ConfirmModalContent';
+import ButtonLink from 'components/admin/ButtonLink/ButtonLink';
 
 import './ActionButtons.scss';
 
 const ActionButtons = ( {
+  id,
   type,
   deleteConfirmOpen,
   setDeleteConfirmOpen,
@@ -77,21 +79,31 @@ const ActionButtons = ( {
         />
       ) }
 
-      { show.preview && (
-        <Modal
-          trigger={ (
-            <Button
-              className="action-btn btn--preview"
-              content="Preview"
-              primary
-              disabled={ disabled.preview }
-            />
-          ) }
-          closeIcon
-        >
-          <Modal.Content>{ previewNode }</Modal.Content>
-        </Modal>
-      ) }
+      { show.preview && type.toLowerCase() !== 'playbook'
+        && (
+          <Modal
+            trigger={ (
+              <Button
+                className="action-btn btn--preview"
+                content="Preview"
+                primary
+                disabled={ disabled.preview }
+              />
+            ) }
+            closeIcon
+          >
+            <Modal.Content>{ previewNode }</Modal.Content>
+          </Modal>
+        ) }
+
+      { show.preview && type.toLowerCase() === 'playbook'
+        && (
+          <ButtonLink
+            content="Preview"
+            disabled={ disabled.preview }
+            url={ `/admin/package/playbook/${id}/preview` }
+          />
+        ) }
 
       { show.publishChanges && (
         <Button
@@ -134,6 +146,7 @@ const ActionButtons = ( {
 };
 
 ActionButtons.defaultProps = {
+  id: '',
   type: 'project',
   deleteConfirmOpen: false,
   setDeleteConfirmOpen: () => {},
@@ -143,6 +156,7 @@ ActionButtons.defaultProps = {
 };
 
 ActionButtons.propTypes = {
+  id: PropTypes.string,
   type: PropTypes.string,
   deleteConfirmOpen: PropTypes.bool,
   setDeleteConfirmOpen: PropTypes.func,